refactor: quebra funções do expander em diferentes arquivos para respeitar a norma42.
